+++ crime_data/common/newmodels.py
@@ -194,58 +194,6 @@ class RetaEstimated(db.Model):
 
     state = relationship(RefState)
 
-class RetaMonthAgencySubcatSummary(db.Model):
-    """
-    Precalculated sums for RETA Agency level data (offense level).
-
-    Create and populate with `dba/create_reta_agency_summary.sql`
-    """
-
-    __tablename__ = 'reta_agency_offense_summary'
-
-    reta_agency_summary_id = db.Column(db.BigInteger, autoincrement=True, primary_key=True)
-    year = db.Column(db.SmallInteger)
-    state_postal_abbr = db.Column(db.Text)
-    state_name = db.Column(db.Text)
-    agency_id = db.Column(db.BigInteger)
-    agency_ori = db.Column(db.Text)
-    agency_name = db.Column(db.Text)
-    reported = db.Column(db.BigInteger)
-    covered = db.Column(db.BigInteger)
-    covering_count = db.Column(db.BigInteger)
-    agency_population = db.Column(db.BigInteger)
-    population_group_code = db.Column(db.Text)
-    population_group = db.Column(db.Text)
-    homicide_reported = db.Column(db.BigInteger)
-    homicide_actual = db.Column(db.BigInteger)
-    homicide_cleared = db.Column(db.BigInteger)
-    homicide_juvenile_cleared = db.Column(db.BigInteger)
-    rape_reported = db.Column(db.BigInteger)
-    rape_actual = db.Column(db.BigInteger)
-    rape_cleared = db.Column(db.BigInteger)
-    rape_juvenile_cleared = db.Column(db.BigInteger)
-
-    def get(self, state = None, agency = None, year = None, county = None):
-        """Get Agency - Offense counts given some filters."""
-        query = RetaMonthAgencySubcatSummary.query
-
-        if state:
-            query = query.filter(func.lower(RetaMonthAgencySubcatSummary.state_postal_abbr) == state.lower())
-        if county:
-            subq = (db.session.query(models.RefAgencyCounty.agency_id)
-                    .select_from(models.RefAgencyCounty)
-                    .join(models.RefCounty, and_(models.RefAgencyCounty.county_id == models.RefCounty.county_id))
-                    .filter(models.RefCounty.county_fips_code == county)
-                )
-            if year:
-                subq = subq.filter(models.RefAgencyCounty.data_year == year)
-            query = query.filter(RetaMonthAgencySubcatSummary.agency_id.in_(subq.subquery()))
-        if agency:
-            query = query.filter(RetaMonthAgencySubcatSummary.agency_ori == agency)
-        if year:
-            query = query.filter(RetaMonthAgencySubcatSummary.year == year)
-
-        return query
 
 class AgencySums(db.Model):
 
@@ -294,6 +242,103 @@ class AgencySums(db.Model):
 
         query = query.join(AgencyParticipation, and_(AgencyParticipation.agency_id == AgencySums.agency_id, AgencyParticipation.year == AgencySums.year)).filter(AgencyParticipation.months_reported == 12)
         query = query.order_by(AgencySums.year.desc()) # Agency reported 12 Months.
+
+        #print(query) # Dubug
+        return query
+
+
+class AgencyOffenseCounts(db.Model):
+    __tablename__ = 'agency_offenses_view'
+
+    id = db.Column(db.BigInteger, autoincrement=True, primary_key=True)
+    year = db.Column(db.SmallInteger)
+    agency_id = db.Column(db.BigInteger)
+    state_postal_abbr = db.Column(db.Text)
+    ori = db.Column(db.Text)
+    pub_agency_name = db.Column(db.Text)
+    offense_id = db.Column(db.BigInteger) # reta_offense_subcat
+    offense_code = db.Column(db.Text) # reta_offense
+    offense_name = db.Column(db.Text)
+    reported = db.Column(db.BigInteger)
+    unfounded = db.Column(db.BigInteger)
+    actual = db.Column(db.BigInteger)
+    cleared = db.Column(db.BigInteger)
+    juvenile_cleared = db.Column(db.BigInteger)
+
+    def get(self, state = None, agency = None, year = None, county = None, explorer_offense = None):
+        """Get Agency Sums given a state/year/county/agency ori, etc."""
+        query = AgencyOffenseCounts.query
+
+        if state:
+            query = query.filter(func.lower(AgencyOffenseCounts.state_postal_abbr) == state.lower())
+        if county:
+            subq = (db.session.query(models.RefAgencyCounty.agency_id)
+                    .select_from(models.RefAgencyCounty)
+                    .join(models.RefCounty, and_(models.RefAgencyCounty.county_id == models.RefCounty.county_id))
+                    .filter(models.RefCounty.county_fips_code == county)
+                )
+            if year:
+                subq = subq.filter(models.RefAgencyCounty.data_year == year)
+            query = query.filter(AgencyOffenseCounts.agency_id.in_(subq.subquery()))
+        if agency:
+            query = query.filter(AgencyOffenseCounts.ori == agency)
+        if year:
+            query = query.filter(AgencyOffenseCounts.year == year)
+        if explorer_offense:
+            offense = ExplorerOffenseMapping(explorer_offense).reta_offense_code
+            query = query.filter(AgencyOffenseCounts.offense_code == offense)
+
+        query = query.join(AgencyParticipation,
+                           and_(AgencyParticipation.agency_id == AgencyOffenseCounts.agency_id,
+                                AgencyParticipation.year == AgencyOffenseCounts.year)).filter(AgencyParticipation.months_reported == 12)
+        query = query.order_by(AgencyOffenseCounts.year.desc()) # Agency reported 12 Months.
+
+        #print(query) # Dubug
+        return query
+
+
+class AgencyClassificationCounts(db.Model):
+    __tablename__ = 'agency_classification_view'
+
+    id = db.Column(db.BigInteger, autoincrement=True, primary_key=True)
+    year = db.Column(db.SmallInteger)
+    agency_id = db.Column(db.BigInteger)
+    state_postal_abbr = db.Column(db.Text)
+    ori = db.Column(db.Text)
+    pub_agency_name = db.Column(db.Text)
+    classification = db.Column(db.Text)
+    reported = db.Column(db.BigInteger)
+    unfounded = db.Column(db.BigInteger)
+    actual = db.Column(db.BigInteger)
+    cleared = db.Column(db.BigInteger)
+    juvenile_cleared = db.Column(db.BigInteger)
+
+    def get(self, state = None, agency = None, year = None, county = None, classification = None):
+        """Get Agency Sums given a state/year/county/agency ori, etc."""
+        query = AgencyClassificationCounts.query
+
+        if state:
+            query = query.filter(func.lower(AgencyClassificationCounts.state_postal_abbr) == state.lower())
+        if county:
+            subq = (db.session.query(models.RefAgencyCounty.agency_id)
+                    .select_from(models.RefAgencyCounty)
+                    .join(models.RefCounty, and_(models.RefAgencyCounty.county_id == models.RefCounty.county_id))
+                    .filter(models.RefCounty.county_fips_code == county)
+                )
+            if year:
+                subq = subq.filter(models.RefAgencyCounty.data_year == year)
+            query = query.filter(AgencyClassificationCounts.agency_id.in_(subq.subquery()))
+        if agency:
+            query = query.filter(AgencyClassificationCounts.ori == agency)
+        if year:
+            query = query.filter(AgencyClassificationCounts.year == year)
+        if classification:
+            query = query.filter(func.lower(AgencyClassificationCounts.classification) == func.lower(classification))
+
+        query = query.join(AgencyParticipation,
+                           and_(AgencyParticipation.agency_id == AgencyClassificationCounts.agency_id,
+                                AgencyParticipation.year == AgencyClassificationCounts.year)).filter(AgencyParticipation.months_reported == 12)
+        query = query.order_by(AgencyClassificationCounts.year.desc()) # Agency reported 12 Months.
 
         #print(query) # Dubug
         return query
@@ -470,6 +515,9 @@ class CdeAgency(db.Model):
     city_name = db.Column(db.String(100))
     state_id = db.Column(db.SmallInteger)
     state_abbr = db.Column(db.String(2))
+    primary_county_id = db.Column(db.BigInteger)
+    primary_county = db.Column(db.String(100))
+    primary_county_fips = db.Column(db.String(5))
     agency_status = db.Column(db.String(1))
     submitting_agency_id = db.Column(db.BigInteger)
     submitting_sai = db.Column(db.String(9))
@@ -493,3 +541,97 @@ class CdeAgency(db.Model):
     staffing_year = db.Column(db.SmallInteger)
     total_officers = db.Column(db.Integer)
     total_civilians = db.Column(db.Integer)
+
+
+class HtAgency(db.Model):
+    """Represents human trafficking counts reported by a single agency in a given year"""
+    class Meta:
+        __tablename__ = 'ht_agency'
+
+    @classmethod
+    def column_is_string(cls, col_name):
+        col = getattr(cls.__table__.c, col_name)
+        return isinstance(col.type, sqltypes.String)
+
+    @classmethod
+    def filtered(cls, filters, args=None):
+        args = args or []
+        qry = cls.query
+
+        # This could be generalized to other places in the future
+        if 'fields' in args:
+            fields = args['fields'].split(',')
+            qry = qry.with_entities(*fields).select_from(CdeAgency)
+
+        for filter in filters:
+            if isinstance(filter, BinaryExpression):
+                qry = qry.filter(filter)
+            else:
+                (col_name, comparitor, values) = filter
+                col = getattr(cls, col_name)
+                if cls.column_is_string(col_name):
+                    col = func.lower(col)
+                operation = getattr(col, comparitor)
+                qry = qry.filter(or_(operation(v) for v in values)).order_by(
+                    col)
+        return qry
+
+    id = db.Column(db.Integer, primary_key=True)
+    year = db.Column(db.SmallInteger)
+    ori = db.Column(db.Text)
+    agency_id = db.Column(db.BigInteger)
+    agency_name = db.Column(db.Text)
+    population = db.Column(db.BigInteger)
+    state_id = db.Column(db.Integer)
+    state_abbr = db.Column(db.Text)
+    months_reported = db.Column(db.SmallInteger)
+    sex_acts = db.Column(db.Integer)
+    sex_acts_cleared = db.Column(db.Integer)
+    sex_acts_juvenile_cleared = db.Column(db.Integer)
+    servitude = db.Column(db.Integer)
+    servitude_cleared = db.Column(db.Integer)
+    servitude_juvenile_cleared = db.Column(db.Integer)
+
+
+class HtSummary(db.Model):
+    """Collects rollups of multiple HtAgency reports. You can use this
+       table to get counts of human trafficking for a given agency, a
+       specific state or national on a single or all years. Note that
+       counts from US Territories are not available in this table (the
+       FBI says they shouldn't be included."""
+    class Meta:
+        __tablename__ = 'ht_summary'
+
+    ht_summary_id = db.Column(db.Integer, primary_key=True)
+    grouping_bitmap = db.Column(db.Integer)
+    year = db.Column(db.SmallInteger)
+    ori = db.Column(db.Text)
+    agency_id = db.Column(db.BigInteger)
+    agency_name = db.Column(db.Text)
+    agencies = db.Column(db.Integer)
+    population = db.Column(db.BigInteger)
+    state_id = db.Column(db.Integer)
+    state_abbr = db.Column(db.Text)
+    months_reported = db.Column(db.SmallInteger)
+    sex_acts = db.Column(db.Integer)
+    sex_acts_cleared = db.Column(db.Integer)
+    sex_acts_juvenile_cleared = db.Column(db.Integer)
+    servitude = db.Column(db.Integer)
+    servitude_cleared = db.Column(db.Integer)
+    servitude_juvenile_cleared = db.Column(db.Integer)
+
+    @classmethod
+    def grouped_by_state(cls, year=None, state_abbr=None):
+        query = HtSummary.query
+
+        query = query.filter(HtSummary.state_id != None)
+        query = query.filter(HtSummary.agency_id == None)
+
+        if year is not None:
+            query = query.filter(HtSummary.year == year)
+
+        if state_abbr is not None:
+            query = query.filter(HtSummary.state_abbr == state_abbr)
+
+        query = query.order_by(HtSummary.year, HtSummary.state_abbr)
+        return query