+++ crime_data/resources/incidents.py
@@ -119,49 +119,54 @@ class AgenciesSumsCounty(CdeResource):
         '''''
         self.verify_api_key(args)
         model = newmodels.AgencySums()
-        year = None
-        if 'year' in args:
-            year = args['year']
-        agency_sums = model.get(agency = agency_ori, year =  year, county = county_fips_code, state=state_abbr)
+        year = args.get('year', None)
+        explorer_offense = args.get('explorer_offense', None)
+        agency_sums = model.get(agency = agency_ori, year =  year, county = county_fips_code, state=state_abbr, explorer_offense=explorer_offense)
         filename = 'agency_sums_county'
         return self.render_response(agency_sums, args, csv_filename=filename)
 
 
-class CachedIncidentsAgenciesCount(CdeResource):
+class AgenciesOffensesCount(CdeResource):
     '''''
     Agency Offense counts by year.
     '''''
-    schema = marshmallow_schemas.CachedAgencyIncidentCountSchema(many=True)
+    schema = marshmallow_schemas.AgencyOffensesSchema(many=True)
     fast_count = True
 
     @use_args(marshmallow_schemas.OffenseCountViewArgs)
     @tuning_page
     def get(self, args, state_abbr = None, agency_ori = None):
         self.verify_api_key(args)
-        model = newmodels.RetaMonthAgencySubcatSummary()
-        year = None
-        if 'year' in args:
-            year = args['year']
-        reta_offenses = model.get(state = state_abbr, agency = agency_ori, year = year)
-        filename = 'agency_sums'
-        return self.render_response(reta_offenses, args, csv_filename=filename)
+        year = args.get('year', None)
+        explorer_offense = args.get('explorer_offense', None)
+        agency_sums = None
+
+        # ugh
+        if explorer_offense == 'violent' or explorer_offense == 'property':
+            agency_sums = newmodels.AgencyClassificationCounts().get(state = state_abbr, agency = agency_ori, year = year, classification = explorer_offense)
+        else:
+            agency_sums = newmodels.AgencyOffenseCounts().get(state = state_abbr, agency = agency_ori, year = year, explorer_offense = explorer_offense)
+        filename = 'agency_offenses_state'
+        return self.render_response(agency_sums, args, csv_filename=filename)
 
 
-class CachedIncidentsAgenciesCountyCount(CdeResource):
+class AgenciesOffensesCountyCount(CdeResource):
     '''''
     Agency Offense counts by year.
     '''''
-    schema = marshmallow_schemas.CachedAgencyIncidentCountSchema(many=True)
+    schema = marshmallow_schemas.AgencyOffensesSchema(many=True)
     fast_count = True
 
-    @use_args(marshmallow_schemas.OffenseCountViewArgs)
+    @use_args(marshmallow_schemas.OffenseCountViewArgsYear)
     @tuning_page
-    def get(self, args, state_abbr = None, agency_ori = None, county_fips_code = None):
+    def get(self, args, state_abbr = None, county_fips_code = None, agency_ori = None):
+        '''''
+        Year is a required field atm.
+        '''''
         self.verify_api_key(args)
-        model = newmodels.RetaMonthAgencySubcatSummary()
-        year = None
-        if 'year' in args:
-            year = args['year']
-        reta_offenses = model.get(state = state_abbr, agency = agency_ori, year = year, county = county_fips_code)
-        filename = 'agency_sums'
-        return self.render_response(reta_offenses, args, csv_filename=filename)
+        model = newmodels.AgencyOffenses()
+        year = args.get('year', None)
+        explorer_offense = args.get('explorer_offense', None)
+        agency_sums = model.get(agency = agency_ori, year =  year, county = county_fips_code, state=state_abbr, explorer_offense=explorer_offense)
+        filename = 'agency_sums_county'
+        return self.render_response(agency_sums, args, csv_filename=filename)