+++ dba/after_load/mat-views/offender-offenses.sql
@@ -85,4 +85,3 @@ create materialized view offense_offender_counts as
 CREATE INDEX offense_offender_counts_state_id_idx ON offense_offender_counts (state_id, year, offense_name);
 CREATE INDEX offense_offender_counts_ori_idx ON offense_offender_counts (ori, year, offense_name);
 
-